refactor(fabrika-cli): move classifyRunSpend + RunSpend into spend/ (#5050)
`classifyRunSpend` is a spend rule, not an eval one: both of its ingredients already
live in `spend/` — it calls `reconstructSpend` from `spend/token-spend.ts`, and its
`RunSpend` return type is a union over that module's `StageSpend`. It named no eval
concept. Living in `eval/runner.ts` made `spend read` import *up* into `eval/`, which
turned `spend ↔ eval` into the only pair of groups in `packages/fabrika-cli/src/` that
import each other, and dragged the corpus schema + grading oracle into `spend read`'s
load path to reach a four-line function.

The function and its type move to `spend/token-spend.ts`, next to the meter they wrap.
Single owner, no re-export shim: every consumer imports from the new home, so the
`eval → spend` edge stays one-way and `spend/` now has zero `../eval/` imports.

Consumers repointed: `eval/runner.ts`, `eval/spawn.ts`, `eval/spawn.unit.test.ts`,
`eval/runner.unit.test.ts`, `spend/read-verb.ts`. The three-arm `RunSpend` unit tests
move with the function into `spend/token-spend.unit.test.ts`. `eval/report.ts`'s own
`RunSpendSchema` is a separate Schema-side union and does not move.

Behavior unchanged — no CLI surface, exit code, or output change.
